@@ -1,14 +1,28 @@
 package com.xworkz.person;
 
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
+import io.swagger.v3.oas.annotations.info.Contact;
+import io.swagger.v3.oas.annotations.info.Info;
 @Configuration
+
+@OpenAPIDefinition(
+		
+		info=@Info(
+				title = "Person API",
+				description = "Doing Crud Operation",
+				summary = "This person api will add,read,update,delete details of a person",
+				contact = @Contact(
+						name = "x-workz",
+						email = "yasmin.xworkz@gmail.com"
+						),
+				version = "v1"
+				
+				
+				)	
+		)
+		
 
 public class SwaggerConfigurationClass {
 	
@@ -16,16 +30,11 @@ public class SwaggerConfigurationClass {
 		System.err.println("running SwaggerConfigurationClass............. ");
 	}
    
-	@Bean
-    public Docket api() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .select()
-                .apis(RequestHandlerSelectors.basePackage("com.xworkz.person.controller"))
-                .paths(PathSelectors.any())
-                .build();
-    }
 
-    
+
+	
+	
+	
    
 }
 
